Implement a multi-step version of the results site
Implement a mock election results site to scrape where the user must
select an election year and then an office through a multi-step form
submitted with an HTTP POST request. The values in the previous step of
the form are passed via hidden form inputs.

In order to provide a space for a link back to the main form page, I
added a new content block to the base template called `aftercontent`.

In order to only show offices for a particular year, I added a `year`
argument to the `get_years_offices()` function.

Addresses #15
https://github.com/asuozzo/nicar2019-scraping/issues/15