Show error page when there are no results
For the version of the page that selects results via URL parameters, the
form is very naive and shows offices for all years. Since an office
might not be up for election in a particular year, there will be no
results. Return an error code of 404 and show the error page when there
aren't any results.

Addresses #18
https://github.com/asuozzo/nicar2019-scraping/issues/18