Add skeletal scrapers for URL parameters form
Implement skeletal scripts to scrape the results form and to fetch the
HTML from the paged results that are retrieved via URL parameters.

Refactor some of the functionality into functions so the place where
participants need to edit the code is more clear and to make it easier
to have skeletal scripts that will run without error, even if the output
will be empty.

To be more robust, check for errors by the request status code not
being 200 rather than checking only for 404 errors.

Addresses #36
https://github.com/asuozzo/nicar2019-scraping/issues/36